requirements.txt
# Semantic Search System - Python 3.10+
# Local-only, no external services

# Embeddings
sentence-transformers>=2.2.0

# Vector store
faiss-cpu>=1.7.4

# ML & clustering
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
